backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
chromadb==0.4.22
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
tiktoken==0.5.2
pytest==7.4.3